moved shared types to utils
