+++ packages/gamut/src/Badge/index.tsx
@@ -3,7 +3,7 @@ import { StyleProps, variance } from '@codecademy/variance';
 import styled from '@emotion/styled';
 
 import { FlexBox } from '../Box';
-import { IconComponentType } from '../Button/shared';
+import { IconComponentType } from '../utils';
 import { determineIconSize, determineIconSpacing } from './helpers';
 
 const colorVariants = variant({
@@ -79,7 +79,9 @@ const BadgeBase = styled('div', styledOptions)<BadgeBaseProps>(
   sizeVariants
 );
 
-export interface BadgeProps extends BadgeBaseProps, Partial<IconComponentType> {}
+export interface BadgeProps
+  extends Partial<IconComponentType>,
+    BadgeBaseProps {}
 
 export const Badge: React.FC<BadgeProps> = ({
   icon: Icon,
