--- packages/gamut/src/utils/childrenType.ts
@@ -1,3 +1,0 @@
-export interface WithChildrenProp {
-  children?: React.ReactNode | React.ReactNode[];
-}
