+++ packages/gamut/src/utils/index.ts
@@ -1,5 +1,5 @@
-export * from './childrenType';
 export * from './createPromise';
 export * from './generateResponsiveClassnames';
 export * from './omitProps';
 export * from './useIsMounted';
+export * from './types';
