+++ packages/styleguide/stories/Atoms/Badge/Badge.stories.mdx
@@ -8,6 +8,7 @@ import { PropsTable } from '@codecademy/storybook-addon-variance';
 import { Canvas, Meta, Story } from '@storybook/addon-docs/blocks';
 import LinkTo from '@storybook/addon-links/react';
 
+import { TertiaryFillExample } from './examples';
 
 <Meta
   title={title}
@@ -64,6 +65,13 @@ Additional variants should designed with both <LinkTo kind='Foundations/ColorMod
   <Story name="accent" args={{ variant: 'accent' }}>
     {BadgeTemplate.bind({})}
   </Story>
+</Canvas>
+
+In the example below that the background is set to `background-primary` (beige).
+Notice that `tertiary` has a transparent background while `tertiaryFill` has a regular `background` color.
+
+<Canvas>
+  <Story name="tertiaryFillSeparate">{() => <TertiaryFillExample />}</Story>
 </Canvas>
 
 ## Size
